codec_adapter: rename the codec folder to module
and fix all the includes accordingly. No functional changes.
