+++ src/audio/codec_adapter/module/passthrough.c
@@ -6,7 +6,7 @@
 //
 // Passthrough codec implementation to demonstrate Codec Adapter API
 
-#include <sof/audio/codec_adapter/codec/generic.h>
+#include <sof/audio/codec_adapter/module/generic.h>
 
 /* 376b5e44-9c82-4ec2-bc83-10ea101afa8f */
 DECLARE_SOF_RT_UUID("passthrough_codec", passthrough_uuid, 0x376b5e44, 0x9c82, 0x4ec2,
