+++ src/include/sof/audio/codec_adapter/module/cadence.h
